Fixed bugs from Itr2
Fixed bug of allowing negative numbers as input for integer values (ex. quantity, price, health card number, phone number, date of birth) so that now it throws an exception.
Fixed bug of healthcard number and phone number being integers and not being able to hold large enough values. Turned these to longs and fixed all associated methods/reading from database/reading from user from int to long.

+++ York and Co. Pharmacy Management System/src/databaseDAO/OrderData/OrderRoot.java
@@ -14,7 +14,7 @@ public interface OrderRoot {
 	public void addToOrderTable(Order o) throws Exception;	
 	public void addToPrescriptionTable (Prescription p) throws Exception;
 	public void addRefillToOrderTable(Order o) throws Exception;	
-	public int numOfRefill (int _patientID, int _medicationId) throws SQLException;	
-	public Boolean checkIfExistsInPrescriptionTable (int _patientID, int _medicationId)  throws SQLException;
+	public int numOfRefill (long _patientID, int _medicationId) throws SQLException;	
+	public Boolean checkIfExistsInPrescriptionTable (long _patientID, int _medicationId)  throws SQLException;
 	public ArrayList<Order> getListOfAllOrders() throws Exception;
 }
