+++ York and Co. Pharmacy Management System/src/databaseDAO/UserData/UserRoot.java
@@ -8,5 +8,5 @@ public interface UserRoot {
 	public ArrayList<User> getListOfUsernamesAndPasswords() throws Exception;
 	public ArrayList<Patient> getListOfAllPatients() throws Exception;
 	public void addPatientToDatabase(Patient newPatient) throws Exception;
-	public void updatePatientInDatabase(int IDOfModifiedPatient, Patient actualPatientObject);
+	public void updatePatientInDatabase(long IDOfModifiedPatient, Patient actualPatientObject);
 }
